+++ security/RELEASE_NOTES.md
@@ -2,6 +2,7 @@
 # Security-sample-template Release Notes
 ## 0.5.0-alpha
 ##### Security
+* SBB-2697 Fixed method return value ```CertificatePinnerCreator.extractPeerCertificate```
 * ANDDEP-687 Changed "com.squareup.okhttp3:logging-interceptor" dependency from "api" to "implementation" type
 ## 0.4.0
 ##### Security