+++ security/RELEASE_NOTES.md
@@ -1,5 +1,8 @@
-[TOC]
 # Security-sample-template Release Notes
+
+- [0.5.0-alpha](#050-alpha)
+- [0.4.0](#040)
+
 ## 0.5.0-alpha
 ##### Security
 * SBB-2697 Fixed method return value ```CertificatePinnerCreator.extractPeerCertificate```